Make sure we are still within 32 bits on Linux numpy

+++ average_steer_partitions.py
@@ -17,8 +17,8 @@ if len(sys.argv) < 2:
 
 def event_to_analog_value(event: ControlEntry):
     val = int32((event.flags << 16) | event.enabled)
-    val <<= 8
-    val >>= 8
+    val <<= int32(8)
+    val >>= int32(8)
     return -val
 
 def partition_steer_events(events: list, sample_period: int):
